+++ src/main/java/nextstep/courses/domain/course/CourseRepository.java
@@ -1,4 +1,6 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.course;
+
+import nextstep.courses.domain.course.Course;
 
 public interface CourseRepository {
     int save(Course course);
