+++ src/main/java/nextstep/courses/domain/session/ImageInfo.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import nextstep.courses.exception.ImageException;
