+++ src/main/java/nextstep/courses/domain/session/Session.java
@@ -1,7 +1,5 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
-import nextstep.courses.exception.PaymentException;
-import nextstep.courses.exception.SessionException;
 import nextstep.payments.domain.Payment;
 import nextstep.users.domain.NsUser;
