+++ src/test/java/nextstep/courses/domain/SessionStateTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.session.SessionState;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
